+++ src/plone.server/plone/server/__init__.py
@@ -5,16 +5,34 @@ import collections
 
 _ = MessageFactory('plone')
 
-DICT_METHODS = {}
-DICT_RENDERS = collections.OrderedDict()
-DICT_LANGUAGES = {}
-DEFAULT_LAYER = []
-DEFAULT_PERMISSION = []
-AVAILABLE_ADDONS = {}
-JSON_API_DEFINITION = {}
-AUTH_EXTRACTION_PLUGINS = []
-AUTH_USER_PLUGINS = []
-CORS = {}
+app_settings = {
+    'databases': [],
+    'address': 8080,
+    'static': [],
+    'utilities': [],
+    'root_user': {
+        'id': 'admin',
+        'password': ''
+    },
+    'auth_policies': [
+        'plone.server.auth.policies.BearerAuthPolicy',
+        'plone.server.auth.policies.WSTokenAuthPolicy',
+    ],
+    'auth_user_identifiers': [
+        'plone.server.auth.users.RootUserIdentifier'
+    ],
+    'auth_token_checker': [
+        'plone.server.auth.checkers.SaltedHashPasswordChecker',
+    ],
+    'default_layers': [],
+    'http_methods': {},
+    'renderers': collections.OrderedDict(),
+    'languages': {},
+    'default_permission': '',
+    'available_addons': {},
+    'api_definition': {},
+    'cors': {}
+}
 
 SCHEMA_CACHE = {}
 PERMISSIONS_CACHE = {}