+++ src/plone.server/plone/server/auth/users.py
@@ -1,0 +1,8 @@
+
+
+class RootUserIdentifier(object):
+    def __init__(self, request):
+        self.request = request
+
+    async def get_user(self):
+        return self.request.application.root_user